requirements.txt
pillow>=12.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
